==> DNNREC_BookCrossing/1. Data Preparation.ipynb <==
PATH='/home/kirana/Documents/phd/experiment1_new/DNNREC_BookCrossing'

DATAPATH='/home/kirana/Documents/final_dissertation_final/experiments/datasets/bookcrossing'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sklearn
import os
import pickle
import torch
from torch.utils.data import Dataset, DataLoader
from torch import nn
import pickle
from sklearn import metrics

ratings=pd.read_csv(f'{DATAPATH}/BX-Book-Ratings.csv',encoding='latin-1',error_bad_lines=False,sep=";")
print (ratings.head())


ratings.shape, ratings['User-ID'].nunique(),ratings['ISBN'].nunique()

ratings.columns=['userId','itemId','rating']

tempdf=ratings[['userId','rating']].groupby('userId').count()
tempdf=tempdf.loc[tempdf['rating']>=20] 
tempdf.shape# Only 7369 users have more than 20 ratings

ratings['num_ratings_per_user_gt_20']=ratings['userId'].isin(tempdf.index)

ratings['num_ratings_per_user_gt_20'].value_counts()

tempdf=ratings[['itemId','rating']].groupby('itemId').count()
tempdf=tempdf.loc[tempdf['rating']>=200] 
tempdf.shape# Only 195 items have more than 20 ratings

ratings['num_ratings_per_item_gt_200']=ratings['itemId'].isin(tempdf.index)

ratings['num_ratings_per_item_gt_200'].value_counts()

# Keep only explicit ratings - implicit means he interacted with the book in some way
ratings=ratings.loc[ratings['rating']>0]

ratings.shape, ratings['userId'].nunique(),ratings['itemId'].nunique()

ratings['flag_clean_considered']=ratings['num_ratings_per_user_gt_20']&ratings['num_ratings_per_item_gt_200']

ratings['flag_clean_considered'].value_counts()

items=pd.read_csv (f'{DATAPATH}/BX-Books.csv',encoding='latin-1',error_bad_lines=False,sep=";")
print(items.head())

items.rename({'ISBN':'itemId'},axis=1,inplace=True)

items.head()

users=pd.read_csv(f'{DATAPATH}/BX-Users.csv',encoding='latin-1',error_bad_lines=False,sep=";")
print (users.head())

users.rename({'User-ID':'userId'},axis=1,inplace=True)

users.head()

ratings.shape

ratings=pd.merge(ratings,users,on='userId',how='left')

ratings.shape

items.columns

ratings=pd.merge(ratings,items.iloc[:,:5],on='itemId',how='left')

ratings.shape

ratings.head()

ratings.shape

dfratings=ratings.copy()

### CROSS-VALIDATION 

# normal method is to keep the first x% of the ratings for each user as training and the rest as validation
# we want to simulate cold start by applying above only to users who were acquired in older years except last 2 years

from sklearn.model_selection import train_test_split

ratings['rating'].value_counts()

users=dfratings['userId'].unique()

users=np.random.permutation(users)

normalfrac=0.9

users_normal=users[:np.int(np.round(len(users)*normalfrac))]

dfnormal=dfratings.loc[dfratings['userId'].isin(users_normal)]

dfsimulate=dfratings.loc[~dfratings['userId'].isin(users_normal)]

dfnormal.shape, dfsimulate.shape

temptrain,tempvalid=train_test_split(dfnormal,test_size=0.1,train_size=0.9,random_state=11,shuffle=True)

temptrain['random_dstype']='train'
tempvalid['random_dstype']='test'

dfnormal=pd.concat([temptrain,tempvalid],axis=0)

dfnormal.shape

dfnormal['random_dstype'].value_counts()

dfsimulate.shape

simulateusers=dfsimulate['userId'].unique()

simulateusers=np.random.permutation(simulateusers)

simulateusers_train=simulateusers[:np.int(np.round(len(simulateusers)*normalfrac))]

dfsimulate_train=dfsimulate.loc[dfsimulate['userId'].isin(simulateusers_train)]
dfsimulate_test=dfsimulate.loc[~dfsimulate['userId'].isin(simulateusers_train)]

dfsimulate_train['random_dstype']='train'
dfsimulate_test['random_dstype']='test'

dfsimulate=pd.concat([dfsimulate_train,dfsimulate_test],axis=0)

dfratings=pd.concat([dfnormal,dfsimulate],axis=0)

dftrain=dfratings.loc[dfratings['random_dstype']=='train']

dfvalid=dfratings.loc[dfratings['random_dstype']!='train']

dftrain.shape, dfvalid.shape

dfvalid['random_dstype'].value_counts()

dfvalid['random_dstype'].value_counts()

dftrain.columns

dftrain.isnull().sum(axis=0)

pickle.dump([dftrain,dfvalid],open(f'{DATAPATH}/df.pkl','wb'))

dftrain['userId'].nunique(),dftrain['itemId'].nunique()
