==> DNNREC_BookCrossing/3. Assess Cold Start.ipynb <==
import pickle

PATH='/home/kirana/Documents/phd/experiment1_new/DNNREC_BookCrossing'
DATAPATH='/home/kirana/Documents/final_dissertation_final/experiments/datasets/bookcrossing'

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import sklearn
import os
import pickle
import torch
from torch.utils.data import Dataset, DataLoader
from torch import nn
import pickle
from sklearn import metrics

[dftrain,dfvalid]=pickle.load(open(f'{DATAPATH}/df_side.pkl','rb'))

trainusers=set(dftrain['userId'].unique())
validusers=set(dfvalid['userId'].unique())
trainitems=set(dftrain['itemId'].unique())
validitems=set(dfvalid['itemId'].unique())

commonusers=trainusers.intersection(validusers)
commonitems=trainitems.intersection(validitems)

dfvalid['cold_start_user']=~dfvalid['userId'].isin(commonusers)
dfvalid['cold_start_item']=~dfvalid['itemId'].isin(commonitems)

print (pd.crosstab(dfvalid['cold_start_user'],dfvalid['cold_start_item']))
print (pd.crosstab(dfvalid['cold_start_user'],dfvalid['cold_start_item'])/dfvalid.shape[0])   

dftrain['userId'].value_counts()

dfvalid['cold_start_group']=np.where(dfvalid['cold_start_user'] & dfvalid['cold_start_item'],'User-Item-cold-start',
                                    np.where((~dfvalid['cold_start_user']) & (~dfvalid['cold_start_item']), 'No-cold-start',\
                                            np.where(dfvalid['cold_start_user'], 'User-Cold-Start', 'Item-Cold_start')))    


dfvalid['cold_start_bucket']=np.where(dfvalid['cold_start_group']=='No-cold-start',0,1)

dfvalid['cold_start_group'].value_counts()

dfvalid['cold_start_bucket'].value_counts()

pickle.dump([dftrain,dfvalid],open(f'{DATAPATH}/df_side_cold.pkl','wb'))

dftrain.shape,dfvalid.shape
